Fix quantity, add show action to /api/invoices

+++ lib/flight/billing/invoice_line_item.ex
@@ -9,7 +9,7 @@ defmodule Flight.Billing.InvoiceLineItem do
   schema "invoice_line_items" do
     field(:rate, :integer)
     field(:amount, :integer)
-    field(:quantity, :integer)
+    field(:quantity, :float)
     field(:description, :string)
     belongs_to(:invoice, Flight.Billing.Invoice)
